third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/add-new-rule-keyboard.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('inspected', next);

  async function next() {
    await Elements.StylesSidebarPane._instance._createNewRuleInViaInspectorStyleSheet();
    eventSender.keyDown("Tab");
    await TestRunner.addSnifferPromise(Elements.StylePropertiesSection.prototype, "_editingSelectorCommittedForTest");

    TestRunner.addResult("Is editing? " + UI.isEditing());
    ElementsTestRunner.dumpSelectedElementStyles(true, false, true);


    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a new rule works properly with user input.
</p>

<div id="inspected">Text</div>

</body>
</html>
